requirements.txt
streamlit==1.28.1
pycryptodome==3.19.0
